Part of #24. Major BLEDevice changes in progress

+++ herald/include/herald/ble/ble_tx_power.h
@@ -8,7 +8,8 @@
 namespace herald {
 namespace ble {
 
-using BLETxPower = int;
+/// \brief Signed short int for Transmit Power in dBm
+using BLETxPower = short int;
 
 } // end namespace
 } // end namespace
